Crate walox_test_util

Test utilities

Re-exports

pub use anyhow;

pub use filecheck;

pub use globwalk;

pub use num_cpus;

pub use regex;

pub use threadpool;

pub use tracing_subscriber;

Modules

filecheck_helpers

Test helper methods that are specific to filecheck tests.

Structs

TestOutput

The result of a named test execution

Enums

TestResult

The outcome of a test.

Traits

Test

A family of tests

Functions

get_panic_message

Take an Any trait object and attempt to cast it to some form of string.

get_workspace_root

Ascending from the current directory, find the workspace root of this project

output_test_outputs

Print a list of TestOutputs and a summary.